fix the backlogged cursors bug
